Added a dry-run mode by default which will roll through a run without action. Running again with the -run argument will act.
Made the scipt print all generated arguments for qemu-system-x86_64 at the end so it's clear what args it'll run with.

Fixed a typo in the vtconsole binding/unbinding function.

Changed the cut's in the PCI/USB enumeration functions to something cleaner.

Added indentations to certain bits in the script so it's easier to read what a dry-run plans to do if run again with -run.

Made the NetworkManager and display-manager killer/reviver more persistent as I sometimes ran into issues.

-bios is automatically populated if left blank, but will now error if the default assumption path it uses is also missing.

General tidy-up of the script.