Short: Added support for tasksetting the VM.
Long:

Updated the README.md for the new -taskset feature and touched up the --help outout in the script.

Added a new flag: -taskset
  This flag takes explicit host threads to execute the VM on. (And adjusts the guest CPU topology respectively)
  I added this as I've played around isolcpus=x,y,z kernel parameter for VM performance.

As an example of this feature.. If I boot with the extra kernel parameter: `isolcpus=0,1,2,3,4,5` it leaves my first 6 threads (3 cores) free of the kernel's scheduling and SMP balancing. I can then specify `-taskset 0,1,2,3` in this script to boot the gaming VM as a 3 core, 2 thread machine while only executing on host thread's 0 to 5 (The first 6). I may add an 'auto' mode for taskset later.

Added "smarter" hyperthreading/smt detection for the default "Pass all threads to guest" default behavior. This was done to incorporate the new -taskset flag.