-image is now optional (good for liveCDs) added -imageformat.
-imageformat can now be specified with -image to tell qemu what the
format of the image you're passing in is going to be, I consider this
better than tacking the ,format=raw on the end of the -image argument.

Both are now also optional -- This is useful for me playing with
passthrough using various liveCD iso's in qemu real quick instead of
worrying about a guest OS just yet. Also useful for a quick Kali+gpu
session which is completely forgotten on VM shutdown.