Typo fixes and sudo changes.
Some text fixups.

Script no longer needs sudo to run and will instead call sudo inside when required. This includes the use of `tee` instead of shell redirection (>).

qemu-system-x86_64 no longer runs as root either and instead as the current user, underprivileged.

chrt and taskset are called on the qemu pid to set priority and host threads to execute on, post-launch.

chown is applied to relevant devices and files before qemu starts for access to the bare minimum.