Add kvm check with helpful troubleshooting output
Had commit f04352397e18c7bcdf2e625f23c623f82cf01f72 from the ubuntu
branch Tue Mar 1 21:55:03 2022 +1100 however it is so far out of date at
this point (And the check so basic) it's not worth porting.

Made this check which responds if /dev/kvm is missing as the script
needs it. If the script can't get the kvm modules going it'll comb dmesg
for pointers to help the user solve the issue when applicable.