Add -nvme. Include storage driver disk info output
-nvme has been added primarily to help deal with P2V scenarios where a
guest which used to be physically installed on NVMe does not have
drivers in their bootloader for using SATA or virtio storage
controllers.

It's also useful for one-off NVMe booting a physical install where an OS
is missing SATA/virtio drivers in the bootloader preventing booting.

Though for long term use I strongly recommend installing the virtio drivers
as soon as possible for any guest.